scanner/requirements.txt
python-dotenv
gitpython
pygithub
flask
gunicorn